You're now able to delete files and if you specify a file that doesn't exist, it'll send back, file doesnt exist, instead of crashing

+++ server.py
@@ -2,6 +2,7 @@
     
 
 import socket
+import os
 import time
 
 OpenedFile = ""
@@ -12,6 +13,7 @@ PORT = 42069
 ADDR = (IP, PORT)
 
 s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+s.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
 s.bind(ADDR)
 
 s.listen(0)
@@ -30,30 +32,52 @@ while True:
 
     match RecievedData[0]:
         case "r":
-            with open(RecievedData[1:], "r") as OpenedFile:
-                conn.send(OpenedFile.read().encode("utf-8"))
-                print("read")
-                OpenedFile.close()
-        
-        case "w":
-            with open(RecievedData[1:], "w") as OpenedFile:
-                conn.send("Send".encode("utf-8"))
-                uwu = ""
-                conn.close()
-                conn, addr = s.accept()
-                RecievedData = conn.recv(1024).decode("utf-8")
+            if os.path.exists(RecievedData[1:]):
+                with open(RecievedData[1:], "r") as OpenedFile:
+                    conn.send(OpenedFile.read().encode("utf-8"))
+                    print("read")
+                    OpenedFile.close()
+            else:
+                conn.send("File doesn't exist".encode("utf-8"))
+                print("File doesn't exist")
                 
-                OpenedFile.write(RecievedData)
-                OpenedFile.close()
-                print("written")
-                conn.send("Written".encode("utf-8"))
+                
+        case "w":
+            if os.path.exists(RecievedData[1:]):
+                with open(RecievedData[1:], "w") as OpenedFile:
+                    conn.send("Send".encode("utf-8"))
+                    uwu = ""
+                    conn.close()
+                    conn, addr = s.accept()
+                    RecievedData = conn.recv(1024).decode("utf-8")
+
+                    OpenedFile.write(RecievedData)
+                    OpenedFile.close()
+                    print("written")
+                    conn.send("Written".encode("utf-8"))
+            else:
+                conn.send("File doesn't exist".encode("utf-8"))
+                print("File doesn't exist")
             
         case "x":
-            with open(RecievedData[1:], "x") as OpenedFile:
-                OpenedFile.close()
-                conn.send("Created".encode("utf-8"))
-                print("created")
+            if os.path.exists(RecievedData[1:]):
+                conn.send("File already exists".encode("utf-8"))
+                print("File already exists")
+            else:
+                with open(RecievedData[1:], "x") as OpenedFile:
+                    OpenedFile.close()
+                    conn.send("Created".encode("utf-8"))
+                    print("created")
         
-        
-        
+        case "f":
+            if os.path.exists(RecievedData[1:]):
+                os.remove(RecievedData[1:])
+                conn.send("Deleted".encode("utf-8"))
+                print("deleted")
+            else:
+                conn.send("File doesn't exist".encode("utf-8"))
+                print("File doesn't exist")
     conn.close()
+    if (RecievedData[0] == "q"):
+        break
+
